LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-anchoring/image-001-expected.txt





FAIL Anchor selection can select images assert_equals: expected 200 but got 150
